+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added endpoint `get_consuming_excess_energy()` to the heat pump
 - Added endpoint `get_excess_energy_target_temperature()` to the buffer tank
 - Added endpoint `set_excess_energy_target_temperature()` to the buffer tank
+- Added endpoint `get_excess_energy_target_temperature_hysteresis()` to the buffer tank
+- Added endpoint `set_excess_energy_target_temperature_hysteresis()` to the buffer tank
 - Added endpoint `get_outdoor_temperature_excess_energy_limit()` to the buffer tank
 - Added endpoint `set_outdoor_temperature_excess_energy_limit()` to the buffer tank
 - Added endpoint `get_use_excess_energy()` to the buffer tank